Control de velocidad por PWM. Se mide la velocidad del motor usando un contador.
Se mantienen componentes que no se utilizarán como speed_DAC o PM1_HA_TIMER

+++ master-tesis/t3sis_v4.cydsn/functions.h
@@ -21,7 +21,7 @@
 extern MOTOR_t PM1;
 
 int32_t fn_mapper(int32_t x, int32_t in_min, int32_t in_max, int32_t out_min, int32_t out_max);
-uint8_t fn_mapper_8b(int32_t x, int32_t in_min, int32_t in_max, int32_t out_min, int32_t out_max);
+uint8_t fn_mapper_8b(int32_t x, int32_t in_min, int32_t in_max, uint8_t out_min, uint8_t out_max);
 
 void ProcessCommandMsg(void);
